# backend/ticket-service/docs/swagger.yaml
basePath: /
definitions:
  domain.AgentStat:
    properties:
      agent_id:
        type: integer
      avg_resolution_seconds:
        type: number
      total_assigned:
        type: integer
      total_resolved:
        type: integer
    type: object
  domain.CriticalTicketRow:
    properties:
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  domain.CriticalTrend:
    properties:
      count:
        type: integer
      tickets:
        items:
          $ref: '#/definitions/domain.CriticalTicketRow'
        type: array
    type: object
  domain.PeriodCount:
    properties:
      count:
        type: integer
      period:
        type: string
      status:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Tickets, comments, attachments and reports as clients see them: through the API Gateway.
    "Try it out" calls the gateway, so the origin of this page must be listed in the gateway ALLOWED_ORIGINS.
    Send the access token from POST /auth/login as `Bearer <token>` in the Authorization header.
  title: Helpdesk Ticket API
  version: "1.0"
paths:
  /reports/agents:
    get:
      description: Admin only. Without `from`/`to` the last 30 days are used.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.AgentStat'
                  type: array
              type: object
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Agent performance
      tags:
      - Reports
  /reports/critical-trends:
    get:
      description: Admin only.
      parameters:
      - default: 24
        description: Time window in hours
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.CriticalTrend'
              type: object
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: High-priority ticket trend
      tags:
      - Reports
  /reports/queue-size:
    get:
      description: Admin only. The data field holds `queue_size`.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Number of unassigned tickets
      tags:
      - Reports
  /reports/summary:
    get:
      description: Admin only. Without `from`/`to` the last 30 days are used.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - default: day
        description: Grouping
        enum:
        - day
        - week
        - month
        - year
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.PeriodCount'
                  type: array
              type: object
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ticket summary per period
      tags:
      - Reports
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"